fix(189pc): FamilyID in FamilyInfoResp struct range overflow(close #6426)

+++ drivers/189pc/types.go
@@ -143,7 +143,7 @@ type FamilyInfoListResp struct {
 type FamilyInfoResp struct {
 	Count      int    `json:"count"`
 	CreateTime string `json:"createTime"`
-	FamilyID   int    `json:"familyId"`
+	FamilyID   int64  `json:"familyId"`
 	RemarkName string `json:"remarkName"`
 	Type       int    `json:"type"`
 	UseFlag    int    `json:"useFlag"`
